Added control so admin can't cut off himself from secmon

+++ controllers/RoleController.php
@@ -101,7 +101,8 @@ class RoleController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if($id != 100)
+            $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
     }
