+++ controllers/UserController.php
@@ -141,7 +141,8 @@ class UserController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if($id != 1)
+            $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
     }
